Fix setup script

+++ setup.sh
@@ -2,7 +2,7 @@
 
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-echo script_dir
+echo ${script_dir}
 
 cd ~/
 
@@ -12,7 +12,9 @@ do
     echo "File '${script_dir}/${file}' does not exist!"
     continue
   fi
-  rm ${file} 2> /dev/null
+  if [ -f ${file} ]; then
+    rm ${file} 2> /dev/null
+  fi
   set -x
   ln -s ${script_dir}/${file} ${file}
   { set +x; } 2>/dev/null
@@ -24,7 +26,9 @@ do
     echo "Directory '${script_dir}/${dir}' does not exist!"
     continue
   fi
-  rm -r ${dir} 2> /dev/null
+  if [ -d ${dir} ]; then
+    rm -r ${dir} 2> /dev/null
+  fi
   set -x
   ln -s ${script_dir}/${dir} ${dir}
   { set +x; } 2>/dev/null
